# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Internship Management API",
        "description": "API for managing student internships and reports",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local server"
        },
        {
            "url": "https://yourdomain.com/api",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/student/reports": {
            "get": {
                "tags": [
                    "Student"
                ],
                "summary": "Get student's reports",
                "description": "Retrieve all reports for the authenticated student",
                "operationId": "88f6179b9d6151dc1239d006f6de9929",
                "responses": {
                    "200": {
                        "description": "Reports retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/StudentReport"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - user is not a student",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/teacher/students-reports": {
            "get": {
                "tags": [
                    "Teacher"
                ],
                "summary": "Get students reports for teacher",
                "description": "Retrieve all student reports for the authenticated teacher with filtering options",
                "operationId": "e4118406e7ae05aaca4f5ba97201f50d",
                "parameters": [
                    {
                        "name": "course",
                        "in": "query",
                        "description": "Filter by course number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "group_id",
                        "in": "query",
                        "description": "Filter by group ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reports retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "reports": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TeacherReport"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - user is not a teacher",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "StudentReport": {
                "properties": {
                    "internship_title": {
                        "type": "string",
                        "example": "Летняя практика"
                    },
                    "teacher": {
                        "type": "string",
                        "example": "Иванов Иван Иванович"
                    },
                    "file_url": {
                        "type": "string",
                        "example": "http://example.com/storage/reports/file.pdf",
                        "nullable": true
                    },
                    "file_name": {
                        "type": "string",
                        "example": "report.pdf"
                    },
                    "grade": {
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "comment": {
                        "type": "string",
                        "example": "Хорошая работа",
                        "nullable": true
                    },
                    "submitted_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-06-15 10:30:00"
                    },
                    "deadline": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-06-30"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-06-01"
                    }
                },
                "type": "object"
            },
            "TeacherReport": {
                "properties": {
                    "report_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "student_fio": {
                        "type": "string",
                        "example": "Иванов Иван Иванович"
                    },
                    "group": {
                        "type": "string",
                        "example": "ПИ-201"
                    },
                    "file_url": {
                        "type": "string",
                        "example": "https://localhost/storage/rewtf.pdf",
                        "nullable": true
                    },
                    "file_name": {
                        "type": "string",
                        "example": "report.pdf"
                    },
                    "grade": {
                        "type": "string",
                        "example": "4",
                        "nullable": true
                    },
                    "comment": {
                        "type": "string",
                        "nullable": true
                    },
                    "submitted_date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deadline": {
                        "type": "string",
                        "format": "date"
                    },
                    "internship_title": {
                        "type": "string",
                        "example": "Летняя практика"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Доступно только для студентов"
                    },
                    "message": {
                        "type": "string",
                        "example": "The provided credentials are incorrect."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}